polyfill para browser completado y pasando los tests

+++ test/browser/test/modules.js
@@ -1,4 +1,4 @@
-window.requireInBrowser = (function(make) {
+window.FooldbBrowserRequire = (function(make) {
     const definitions = {};
     const definitionsByOrder = [];
     const defineModule = function(id, value) {
@@ -247,37 +247,31 @@ window.requireInBrowser = (function(make) {
         __DEFINE_MODULE__("/test/features/0010. Operaciones principales/0010. Insertar datos.test.js", exportation);
     })(function(module) {
         module.exports = async fooldb => {
-            console.log("swing 1");
             const Persona1 = await fooldb.insert("Persona", {
                 "nombre": "Persona1",
                 "edad": 35,
                 "fecha de nacimiento": "1991/01/05"
             });
-            console.log("swing 2");
             const Persona2 = await fooldb.insert("Persona", {
                 "nombre": "Persona2",
                 "edad": 35,
                 "fecha de nacimiento": "1991/01/10"
             });
-            console.log("swing 3");
             const Persona3 = await fooldb.insert("Persona", {
                 "nombre": "Persona3",
                 "edad": 35,
                 "fecha de nacimiento": "1991/01/15"
             });
-            console.log("swing 4");
             const Lugar1 = await fooldb.insert("Lugar", {
                 "nombre": "Lugar1",
                 "presidente": Persona1,
                 "habitantes": [Persona1, Persona2, Persona3],
             });
-            console.log("swing 5");
             const Lugar2 = await fooldb.insert("Lugar", {
                 "nombre": "Lugar2",
                 "presidente": Persona1,
                 "habitantes": [Persona1, Persona2, Persona3],
             });
-            console.log("swing 6");
         };
     });
     (function(factory) {
@@ -297,7 +291,7 @@ window.requireInBrowser = (function(make) {
         const modulez = {};
         factory(modulez);
         const exportation = modulez.exports;
-        __DEFINE_MODULE__("/test/features/0010. Operaciones principales/0020.Actualizar datos.test.js", exportation);
+        __DEFINE_MODULE__("/test/features/0010. Operaciones principales/0020. Actualizar datos.test.js", exportation);
     })(function(module) {
         module.exports = async fooldb => {
             await fooldb.update("Persona", row => row.nombre === "Persona2", {
@@ -316,7 +310,7 @@ window.requireInBrowser = (function(make) {
         const modulez = {};
         factory(modulez);
         const exportation = modulez.exports;
-        __DEFINE_MODULE__("/test/features/0010. Operaciones principales/0025.Eliminar datos.test.js", exportation);
+        __DEFINE_MODULE__("/test/features/0010. Operaciones principales/0025. Eliminar datos.test.js", exportation);
     })(function(module) {
         module.exports = async fooldb => {
             await fooldb.delete("Persona", row => row.nombre === "Persona2 modificada");
@@ -886,9 +880,13 @@ window.requireInBrowser = (function(make) {
             fooldb.isArrayOfIntegers("no");
             fooldb.composePath("whatever");
             fooldb.trace.activate();
-            await fooldb.readJson(__dirname + "/../../../package.json");
-            await fooldb.writeJson(__dirname + "/temporary.json", {});
-            await fooldb.constructor.fs.promises.unlink(__dirname + "/temporary.json");
+            Solo_en_nodejs: {
+                if (fooldb.constructor.runningOn.nodejsOnly) {
+                    await fooldb.readJson(__dirname + "/../../../package.json");
+                    await fooldb.writeJson(__dirname + "/temporary.json", {});
+                    await fooldb.constructor.fs.promises.unlink(__dirname + "/temporary.json");
+                }
+            }
             try {
                 await fooldb.insert("Lugar", {
                     nombre: "whatever",
